feat: patch bower.json when when it does not exist
When `bower.json` file does not exist, package installation
may fail because of the missing configuration.
Since `package.json` is pretty close to the `bower.json`, we
try to patch the package.
For now, `dependencies` and `devDependencies` are ignored since
it may not be resolved correctly (should we prefix each
dependencies with the bower-npm-resolver prefix ?).